rename operatorversions
