+++ service/controller/resource/certconfig/desired.go
@@ -29,8 +29,8 @@ func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) (interf
 
 	// When the CertConfig CR is deleted we do not need to compute the desired
 	// state, because we only use the current state to delete the CR. Also note
-	// that the desired state relies on the operatorversions resource, because we
-	// put the cert-operator version into the CR. The operatorversions resource
+	// that the desired state relies on the releaseversions resource, because we
+	// put the cert-operator version into the CR. The releaseversions resource
 	// does not fill the controller context with versions on delete events, which
 	// is also why we cannot compute the correct desired state. We do not want to
 	// fetch the version information on delete events to reduce eventual friction.
