+++ service/controller/resource/releaseversions/create.go
@@ -1,4 +1,4 @@
-package operatorversions
+package releaseversions
 
 import (
 	"context"
